split <Data> into self-contained components for each section
