+++ src/components/Data/d3/drawFruitHist.js
